// src/crawlers/coursera/course-list.interface.ts
export interface CourseList {
    avgLearningHours: any;
    tagline: any;
    name: string;
    partnerLogos: string[],
    isPartOfCourseraPlus: boolean;
    subtitleLanguage: string[];
    productDifficultyLevel: string;
    audience: any;
    enrollments: number;
    partners: string[];
    language: string;
    entityType: string;
    numProductRatings: number;
    allLanguageCodes: string[];
    avgProductRating: number;
    skills: string[];
    imageUrl: string;
    isCourseFree: boolean;
    objectUrl: string;
    objectID: string;
    _snippetResult: any,
    _highlightResult: any
}